# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TeamTime API",
    "version": "1.0.0",
    "description": "API REST para gestión de tiempo, proyectos y recursos humanos con sistema de roles jerárquico",
    "contact": {
      "name": "TeamTime Development Team",
      "email": "desarrollo@teamtime.com",
      "url": "https://github.com/your-org/teamtime-backend"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.teamtime.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Descripción del error"
          },
          "error": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "properties": {
              "field": {
                "type": "string",
                "example": "email"
              },
              "code": {
                "type": "string",
                "example": "INVALID_FORMAT"
              }
            }
          }
        }
      },
...
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "projectId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "task-123"
          },
          "title": {
            "type": "string",
            "example": "Implementar autenticación JWT"
          },
          "description": {
            "type": "string",
            "example": "Desarrollar sistema de autenticación con tokens JWT"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "example": "project-456"
          },
          "assignedUserId": {
            "type": "string",
            "format": "uuid",
            "example": "user-789"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDIENTE",
              "EN_PROGRESO",
              "COMPLETADA",
              "CANCELADA"
            ],
            "example": "EN_PROGRESO"
          },
          "priority": {
            "type": "string",
            "enum": [
              "BAJA",
              "MEDIA",
              "ALTA",
              "URGENTE"
            ],
            "example": "ALTA"
          },
          "estimatedHours": {
            "type": "number",
            "format": "float",
            "example": 8.5
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "example": "2025-07-10"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-04T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-04T10:30:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Token de autenticación inválido o faltante",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Token de autenticación requerido",
              "error": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Permisos insuficientes para realizar la operación",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "No tienes permisos para realizar esta acción",
              "error": "FORBIDDEN"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Recurso no encontrado",
              "error": "NOT_FOUND"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Error de validación en los datos enviados",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Error de validación",
              "error": "VALIDATION_ERROR",
              "details": {
                "field": "email",
                "code": "INVALID_FORMAT"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tasks": {
      "post": {
        "summary": "Crear nueva tarea",
        "description": "Permite a administradores y coordinadores crear una nueva tarea en un proyecto.",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "projectId"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Implementar autenticación JWT"
                  },
                  "description": {
                    "type": "string",
                    "example": "Desarrollar sistema de autenticación con tokens JWT"
                  },
                  "projectId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "project-123"
                  },
                  "assignedUserId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "user-456"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "BAJA",
                      "MEDIA",
                      "ALTA",
                      "URGENTE"
                    ],
                    "default": "MEDIA",
                    "example": "ALTA"
                  },
                  "estimatedHours": {
                    "type": "number",
                    "format": "float",
                    "minimum": 0.1,
                    "example": 8.5
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-07-15"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDIENTE",
                      "EN_PROGRESO",
                      "COMPLETADA",
                      "CANCELADA"
                    ],
                    "default": "PENDIENTE",
                    "example": "PENDIENTE"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tarea creada exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "get": {
        "summary": "Obtener lista de tareas",
        "description": "Retorna una lista paginada de tareas con filtros opcionales. Los usuarios ven tareas según sus permisos.",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Cantidad de elementos por página"
          },
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por proyecto"
          },
          {
            "in": "query",
            "name": "assignedUserId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por usuario asignado"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "EN_PROGRESO",
                "COMPLETADA",
                "CANCELADA"
              ]
            },
            "description": "Filtrar por estado de la tarea"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string",
              "enum": [
                "BAJA",
                "MEDIA",
                "ALTA",
                "URGENTE"
              ]
            },
            "description": "Filtrar por prioridad"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por título o descripción"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tareas obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tareas obtenidas exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tasks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Task"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "limit": {
                              "type": "integer",
                              "example": 10
                            },
                            "total": {
                              "type": "integer",
                              "example": 85
                            },
                            "pages": {
                              "type": "integer",
                              "example": 9
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Obtener tarea por ID",
        "description": "Retorna la información detallada de una tarea específica.",
        "tags": [
          "Tareas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID de la tarea"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tarea obtenida exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  }
}